# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Ocytosign",
        "description": "Documentation de l'API Ocytosign",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/rdvs": {
            "get": {
                "tags": [
                    "RDV"
                ],
                "summary": "Liste tous les RDVs",
                "operationId": "5ca4607af4c564ea3db6ce2faeb3306b",
                "responses": {
                    "200": {
                        "description": "Liste des RDVs récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "RDV"
                ],
                "summary": "Créer un nouveau RDV",
                "operationId": "8f9824ccabe6299d177ecc540b2891b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_client",
                                    "id_medecin",
                                    "date_debut",
                                    "date_fin",
                                    "presentiel"
                                ],
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_medecin": {
                                        "type": "integer"
                                    },
                                    "id_interprete": {
                                        "type": "integer"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "presentiel": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "RDV créé avec succès"
                    }
                }
            }
        },
        "/api/rdvs/{id}": {
            "get": {
                "tags": [
                    "RDV"
                ],
                "summary": "Afficher un RDV précis",
                "operationId": "285323a84483cf7f373a15102171cdc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RDV trouvé"
                    },
                    "404": {
                        "description": "RDV non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "RDV"
                ],
                "summary": "Mettre à jour un RDV",
                "operationId": "13282cb588f49680d66523a55160014a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_medecin": {
                                        "type": "integer"
                                    },
                                    "id_interprete": {
                                        "type": "integer"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "presentiel": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RDV mis à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "RDV"
                ],
                "summary": "Supprimer un RDV",
                "operationId": "ab61943016968bece2811947693e7858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RDV supprimé avec succès"
                    },
                    "404": {
                        "description": "RDV non trouvé"
                    }
                }
            }
        }
    },
...
}